organized codebase into subpackages, fixed chrome import paths
