+++ pkg/chrome/chrome.go
@@ -1,14 +1,15 @@
-package kooky
+package chrome
 
 import (
 	"fmt"
 	"time"
 
 	"github.com/go-sqlite/sqlite3"
+	kooky "github.com/kgoins/kooky/pkg"
 )
 
-func ReadChromeCookies(filename string, domainFilter string, nameFilter string, expireAfter time.Time) ([]*Cookie, error) {
-	var cookies []*Cookie
+func ReadChromeCookies(filename string, domainFilter string, nameFilter string, expireAfter time.Time) ([]*kooky.Cookie, error) {
+	var cookies []*kooky.Cookie
 	db, err := sqlite3.Open(filename)
 	if err != nil {
 		return nil, err
@@ -19,7 +20,7 @@ func ReadChromeCookies(filename string, domainFilter string, nameFilter string, 
 		if rowId == nil {
 			return fmt.Errorf("unexpected nil RowID in Chrome sqlite database")
 		}
-		cookie := &Cookie{}
+		cookie := &kooky.Cookie{}
 
 		// TODO(zellyn): handle older, shorter rows?
 		if len(rec.Values) < 14 {
