+++ pkg/chrome/chrome_test.go
@@ -1,4 +1,4 @@
-package kooky
+package chrome
 
 import (
 	"testing"
