+++ pkg/kooky.go
